Исправление после ревью
