--- packages/client/src/components/App/App.tsx
@@ -1,18 +1,0 @@
-import { useEffect } from 'react'
-import './App.css'
-
-function App() {
-  useEffect(() => {
-    const fetchServerData = async () => {
-      const url = `http://localhost:${__SERVER_PORT__}`
-      const response = await fetch(url)
-      const data = await response.json()
-      console.log(data)
-    }
-
-    fetchServerData()
-  }, [])
-  return <div className="App">Вот тут будет жить ваше приложение! :)</div>
-}
-
-export default App
