+++ packages/client/src/pages/home/index.tsx
@@ -1,8 +1,17 @@
-import { FC } from 'react'
-import App from '../../components/App/App'
+import { FC, useEffect } from 'react'
 
 const Home: FC = () => {
-  return <App />
+  useEffect(() => {
+    const fetchServerData = async () => {
+      const url = `http://localhost:${__SERVER_PORT__}`
+      const response = await fetch(url)
+      const data = await response.json()
+      console.log(data)
+    }
+
+    fetchServerData()
+  }, [])
+  return <div className="App">Вот тут будет жить ваше приложение! :)</div>
 }
 
 export default Home
